No more six

+++ 0x0F-python-object_relational_mapping/model_state.py
@@ -16,7 +16,7 @@ class State(Base):
     It inherits from the Base class
     """
     __tablename__ = "states"
-    id = Column("id", Integer, primary_key=True, nullable=False,
+    id = Column("id", Integer(11), primary_key=True, nullable=False,
                 autoincrement=True)
     name = Column("name", String(128), nullable=False)
